Move audit to InitEntropy()
- InitEntropy is in `GenericPlatformManagerImpl` and should always
  be called. Lack of the logs will help catch bad init.
