+++ src/platform/Entropy.cpp
@@ -23,8 +23,60 @@
  */
 
 #include <crypto/CHIPCryptoPAL.h>
+#include <lib/support/CodeUtils.h>
+#include <lib/support/RandUtils.h>
+
+// Temporary includes for TemporaryAuditRandomPerformance()
+// TODO: remove once https://github.com/project-chip/connectedhomeip/issues/10454 is done.
+#if 1
+#include <lib/support/BytesToHex.h>
+#endif
 
 namespace chip {
+
+namespace {
+
+// Audit random number generator proper initialization with prints.
+// TODO: remove once https://github.com/project-chip/connectedhomeip/issues/10454 is done.
+void TemporaryAuditRandomNumberGenerator()
+{
+    uint8_t buf1[16];
+    uint8_t buf2[16];
+
+    memset(&buf1[0], 0, sizeof(buf1));
+    memset(&buf2[0], 0, sizeof(buf2));
+
+    VerifyOrDie(chip::Crypto::DRBG_get_bytes(&buf1[0], sizeof(buf1)) == CHIP_NO_ERROR);
+    VerifyOrDie(chip::Crypto::DRBG_get_bytes(&buf2[0], sizeof(buf2)) == CHIP_NO_ERROR);
+
+    char hex_buf[sizeof(buf1) * 2 + 1];
+
+    ChipLogProgress(DeviceLayer, "AUDIT: ===== RANDOM NUMBER GENERATOR AUDIT START ====");
+    ChipLogProgress(DeviceLayer, "AUDIT: * Validate buf1 and buf2 are <<<different every run/boot!>>>");
+    ChipLogProgress(DeviceLayer, "AUDIT: * Validate r1 and r2 are <<<different every run/boot!>>>");
+
+    memset(&hex_buf[0], 0, sizeof(hex_buf));
+    VerifyOrDie(Encoding::BytesToUppercaseHexString(&buf1[0], sizeof(buf1), &hex_buf[0], sizeof(hex_buf)) == CHIP_NO_ERROR);
+    ChipLogProgress(DeviceLayer, "AUDIT: * buf1: %s", &hex_buf[0]);
+
+    memset(&hex_buf[0], 0, sizeof(hex_buf));
+    VerifyOrDie(Encoding::BytesToUppercaseHexString(&buf2[0], sizeof(buf2), &hex_buf[0], sizeof(hex_buf)) == CHIP_NO_ERROR);
+    ChipLogProgress(DeviceLayer, "AUDIT: * buf2: %s", &hex_buf[0]);
+
+    VerifyOrDieWithMsg(memcmp(&buf1[0], &buf2[0], sizeof(buf1)) != 0, DeviceLayer,
+                       "AUDIT: FAILED: buf1, buf2 are equal: DRBG_get_bytes() does not function!");
+
+    uint32_t r1 = GetRandU32();
+    uint32_t r2 = GetRandU32();
+
+    ChipLogProgress(DeviceLayer, "AUDIT: * r1: 0x%08" PRIX32 " r2: 0x%08" PRIX32, r1, r2);
+    VerifyOrDieWithMsg(r1 != r2, DeviceLayer,
+                       "AUDIT: FAILED: buf1, buf2 are equal: random number generator does not function!");
+    ChipLogProgress(DeviceLayer, "AUDIT: ===== RANDOM NUMBER GENERATOR AUDIT END ====");
+}
+
+} // namespace
+
 namespace DeviceLayer {
 namespace Internal {
 
@@ -33,6 +85,11 @@ CHIP_ERROR InitEntropy()
     unsigned int seed;
     ReturnErrorOnFailure(chip::Crypto::DRBG_get_bytes((uint8_t *) &seed, sizeof(seed)));
     srand(seed);
+
+    {
+        // TODO: remove once https://github.com/project-chip/connectedhomeip/issues/10454 is done.
+        TemporaryAuditRandomNumberGenerator();
+    }
     return CHIP_NO_ERROR;
 }
